feat: Add build script for generating UMD and minified files
This commit adds a new build script for generating UMD and minified files. The script uses the 'unbuild' package to build the files. It includes configurations for rollup, including options for generating UMD format, minifying the code, and copying files to the appropriate directories. The script also includes hooks for performing additional tasks after the build is done.