feat: Add removeItem script
This commit adds the `removeItem` script, which is responsible for deleting files and directories. It uses the `fs.rm` method to recursively delete items and logs a message when an item is successfully deleted. The script takes an input object with `rootPath` and `tmps` properties. If `rootPath` is provided, it searches for `.tgz` files in the specified directory and deletes them. If `tmps` is provided, it deletes each item in the array.